setup all external service provider instances in geoCoder
This is done by going through the services.json file
The file is also checked for consistency with the services.secrets.json
file for secrets information.

The primary provider is respected by putting that provider at the top.
If the primary provider is not found in the secrets.json file, a
ValueErorr is thrown.