Update for primary service provider choice
Choice of primary service provider is implemented through the
services.json file. If "primary" field of a provider is true, it is used
as the prirmary choice. If none have "primary" field to be true, any one
of providers can be chosen as primary. if many have "primary" field to
be true, one of them at random is chosen as the primary provider.